requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.1.4
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
beautifulsoup4==4.12.2
